Fixed the 4.4 cases and added new cases

+++ src/sync.sh
@@ -9,6 +9,7 @@ helpFunction()
    echo -e "\t-e | --environment: The environment that tells the script will load variables in the file \"settings.<env>\". (*production|staging)"
    echo -e "\t-r | --syslog: The option indicator that defines whether enables target system to allow to receive syslog by 514 port or not. (*disable|enable)"
    echo -e "\t-u | --upgrade: The option indicator that defines whether upgrade packages or not (*disable|enable)"
+   echo -e "\t-k | --keep-firewall: The option indicator that defines whether keep the firewall rules or not (*no|yes)"
    exit 0 # Exit script after printing help
 }
 
@@ -17,8 +18,9 @@ RECV_SYSLOG="disable"
 ENVIR="production"
 FORWARD="enable"
 UPGRADE="disable"
+KEEP_FIREWALL="no"
 
-PARSED_ARGUMENTS=$(getopt -o s:e:ru --long upgrade,syslog,server:,environment: -- "$@")
+PARSED_ARGUMENTS=$(getopt -o s:e:ruk --long upgrade,syslog,keep-firewall,server:,environment: -- "$@")
 
 while true; do
     case "$1" in
@@ -26,6 +28,7 @@ while true; do
         -e|--environment ) ENVIR="$2"; shift 2 ;;
         -u|--upgrade ) UPGRADE="enable"; shift 1;;
         -r|--syslog ) FORWARD="disable"; RECV_SYSLOG="enable"; shift 1;;
+        -k|--keep-firewall) KEEP_FIREWALL="yes"; shift 1;;
         --|"" ) shift; break ;;
         * ) helpFunction;  break ;; # Print helpFunction in case parameter is non-existent
     esac
@@ -61,8 +64,14 @@ if [ "$UPGRADE" != "disable" -a "$UPGRADE" != "enable" ]; then
     exit 1
 fi
 
+if [ "$KEEP_FIREWALL" != "yes" -a "$KEEP_FIREWALL" != "no" ]; then
+    printf '%s\n' "Invalid keep firewall option. (yes|no)" >&2
+    helpFunction
+    exit 1
+fi
+
 echo "Copying the required file into the /root path"
 scp -qr cis/ root@$SRV_IP:/root
 echo "Start run the script with the environment \"$ENVIR\"..."
-ssh root@$SRV_IP bash -c "echo $ENVIR && echo $RECV_SYSLOG && echo $FORWARD && echo $UPGRADE && /root/cis/run.sh $ENVIR $RECV_SYSLOG $FORWARD $UPGRADE"
+ssh root@$SRV_IP ENVIR=$ENVIR RECV_SYSLOG=$RECV_SYSLOG FORWARD=$FORWARD UPGRADE=$UPGRADE KEEP_FIREWALL=$KEEP_FIREWALL /root/cis/run.sh
 exit 0
